refactor: CenterStatus 패키지 변경
