+++ src/main/java/com/trustamarket/inspectionservice/center/domain/enums/CenterStatus.java
@@ -1,7 +1,32 @@
 package com.trustamarket.inspectionservice.center.domain.enums;
 
+import java.util.EnumSet;
+import java.util.Set;
+
 public enum CenterStatus {
-    OPEN,
-    MAINTENANCE,
-    CLOSED
+
+    OPEN {
+        @Override
+        public Set<CenterStatus> allowedTransitions() {
+            return EnumSet.of(MAINTENANCE);
+        }
+    },
+    MAINTENANCE {
+        @Override
+        public Set<CenterStatus> allowedTransitions() {
+            return EnumSet.of(OPEN, CLOSED);
+        }
+    },
+    CLOSED {
+        @Override
+        public Set<CenterStatus> allowedTransitions() {
+            return EnumSet.noneOf(CenterStatus.class);
+        }
+    };
+
+    public abstract Set<CenterStatus> allowedTransitions();
+
+    public boolean canTransitionTo(CenterStatus target) {
+        return allowedTransitions().contains(target);
+    }
 }
