+++ src/test/java/com/trustamarket/inspectionservice/center/application/service/InspectionCenterServiceTest.java
@@ -4,7 +4,7 @@ import com.trustamarket.inspectionservice.center.application.dto.command.Registe
 import com.trustamarket.inspectionservice.center.application.dto.result.ChangeCenterStatusResult;
 import com.trustamarket.inspectionservice.center.application.dto.result.RegisterCenterResult;
 import com.trustamarket.inspectionservice.center.application.port.out.InspectionCenterRepository;
-import com.trustamarket.inspectionservice.center.domain.model.enums.CenterStatus;
+import com.trustamarket.inspectionservice.center.domain.enums.CenterStatus;
 import com.trustamarket.inspectionservice.center.domain.exception.InspectionCenterException;
 import com.trustamarket.inspectionservice.center.domain.model.InspectionCenter;
 import com.trustamarket.inspectionservice.center.domain.vo.Address;
